common: rename common/external to common/online
"External" said where the code was not rather than what it did, and it sat next
to common/external_service.py, which is a local dynamic-import helper - so the
package and the module that looked like it belonged to it meant opposite things.

"Remote" was the obvious replacement and is worse: remote means the Remote
Control page throughout this repo. "Online" says it plainly and collides with
nothing.

common/host keeps its name on purpose. It is the same word as
RESIDENCY_PLAY_HOST in the API, so one concept has one name across the contract,
the capability list and the file layout.
