--- common/external/__init__.py
@@ -1,1 +1,0 @@
-"""Services outside this machine: VPSdb, the theme registry, updates."""
