+++ common/online/__init__.py
@@ -1,0 +1,1 @@
+"""Services reached over the internet: VPSdb, the theme registry, updates."""
