+++ common/online/theme_registry_client.py
@@ -9,7 +9,7 @@ import requests
 from common.http_client import get_json
 
 
-logger = logging.getLogger("vpinfe.common.external.theme_registry_client")
+logger = logging.getLogger("vpinfe.common.online.theme_registry_client")
 
 
 class ThemeRegistryError(Exception):
