+++ frontend/runtime.py
@@ -13,7 +13,7 @@ from frontend.ws_bridge import WebSocketBridge
 from common.host import system_actions
 from common.config_access import DisplayConfig, NetworkConfig, SettingsConfig
 from common.host.display_service import get_display_monitors
-from common.external.vpinplay_runtime import clear_alternate_profile
+from common.online.vpinplay_runtime import clear_alternate_profile
 
 
 WINDOW_CONFIGS = (
